+++ src/slogo/model/movementcommand/Backward.java
@@ -43,7 +43,7 @@ public class Backward extends MovementCommand implements Command {
    * @return number of pixels moved
    */
   @Override
-  public int execute(Turtle turtle) {
+  public double execute(Turtle turtle) {
     turtle.setOrientation(turtle.getOrientation() + 180);
     turtle.setLocation(
         determineLocation(turtle.getLocation(), Math.toRadians(turtle.getOrientation()), pixels));
